FDS Source: rewrite test_filter (tophat)
User a 3x3x3 array as a kernel. Might be advantageous to explicitely
write this as several loops instead. Need to measure serial performance
to see if it is worth the loss of legibility.

The treatment of edges is based on the previous handling and by no means
perfect. But it apparently was sufficient previously so it will
have to do.

Removes the min max checking for the edges. Usually 1e10 and -1e10 were
passed. This seems suppurfluous to me.


Former-commit-id: b5aedf8b185111f4ce4af076c8e8c3dce2e0bc6b [formerly cc905031a570e3e008f19068997826ea4f024549]
Former-commit-id: 11b869cd049ce8d3bffbb79b978b0bb807c31f17