tkorhon1: FDS Source, evacuation time change in EVAC_MAIN_LOOP.
Added line in EVAC_MAIN_LOOP so that at ICYC=1 T_EVAC is a little bit
less than T_FIRE. This way the EVAC_EFF file is written at ICYC=1.
This used to be the case before (when each fire mesh nad differend DT).
This was broken in the process, where same T and DT was made for all
fire meshes.


Former-commit-id: 68e63a9dcdd8269a867c4b2ef65bc882e92dd706 [formerly 15791a852b3dc22c965e1eb8cf85d570b913abbf]
Former-commit-id: 750ad4010366aec3ef2e9e7651b1ac49e20eebbf